Refactor CLI tests
