+++ tests/integration/cli/test_cli_entrypoint.py
@@ -1,0 +1,28 @@
+import sys
+from unittest.mock import patch
+
+import pytest
+
+from src.main import main
+
+pytestmark = pytest.mark.integration
+
+
+def test_main_when_main(sample_config_files):
+    env_path, model_path, trainer_path = sample_config_files
+    test_args = [
+        "main.py",
+        "--env-config",
+        env_path,
+        "--model-config",
+        model_path,
+        "--trainer-config",
+        trainer_path,
+    ]
+    with (
+        patch.object(sys, "argv", test_args),
+        patch("src.main.Trainer") as mock_trainer_class,
+        patch("src.main.main") as mock_main,
+    ):
+        exec(open("src/main.py").read())
+        assert callable(main)
